all basic tests passed, 13 failed Mouli

+++ sources/flag_o.c
@@ -6,7 +6,7 @@
 /*   By: aschukin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/14 16:14:20 by aschukin          #+#    #+#             */
-/*   Updated: 2018/04/02 18:18:14 by aschukin         ###   ########.fr       */
+/*   Updated: 2018/04/03 17:13:10 by aschukin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,8 @@ static uintmax_t	length_o(va_list *ap, t_print *arg)
 	else if (arg->length == 7)
 		return ((size_t)va_arg(*ap, uintmax_t));
 	else
-	//	return ((unsigned int)va_arg(*ap, uintmax_t));
-		return (va_arg(*ap, uintmax_t));
+		return ((unsigned int)va_arg(*ap, uintmax_t));
+	//	return (va_arg(*ap, uintmax_t));
 }
